v1.0 - bugsfixing upload image di post traits: done

+++ src/appbase/arins/Traits/Http/Controller/Post/Action.php
@@ -33,8 +33,6 @@ trait Action
 
         //validate input value
         $validator = Validator::make($data, $this->data->getValidateInput());
-        //return dd($validator->attributes());
-        //return dd($validator->errors());
         if ($validator->fails()) {
             //step 2: Kembali ke halaman input
             return redirect()->route($this->sViewName . '.create')
@@ -72,6 +70,7 @@ trait Action
         $upload = $request->file('upload'); //upload file (image/document) ==> if included
         $imageTemp = $request->input('imageTemp'); //temporary file uploaded
         $toggleRemoveImage = $request->input('toggleRemoveImage');
+        //return dd($toggleRemoveImage);
 
         // return dd($data);
         // //convert input value (string/date/number/email/etc)
@@ -85,18 +84,35 @@ trait Action
         //validate input value
         $request->validate($this->data->getValidateInput());
 
-        //copy temporary uploaded image to real path
-        $data['image'] = Filex::uploadOrCopyAndRemove('', $uploadTemp, 'activities', $upload, 'public', false);
-        // Filex::delete($imageOld);
+        $imageNew = Filex::uploadOrCopyAndRemove($imageOld, $uploadTemp, 'activities', $upload, 'public', false);
+        $data['image'] = $imageNew;
+        if (strtolower($toggleRemoveImage) ==  'true')
+        {
+            $data['image'] = null;
+        }
 
         if ($this->data->update($record, $data)) {
-            Filex::delete($imageOld);
+            if ($uploadTemp != null)
+            {
+                Filex::delete($imageOld);
+                Filex::delete($uploadTemp);
+            } //end if
+
+            if (strtolower($toggleRemoveImage) == 'true')
+            {
+                Filex::delete($imageOld);
+                Filex::delete($imageNew);
+                Filex::delete($uploadTemp);
+            }
             return redirect()->route($this->sViewName . '.index');
         }
 
         /** jika tetap terjadi kesalahan maka ada kesalahan pada system */
         //step 1: delete image if fail to save
-        Filex::delete($data['image']);
+        if ($uploadTemp != null)
+        {
+            Filex::delete($data['image']);
+        } //end if
 
         //step 2: Kembali ke halaman input
         return redirect()->route($this->sViewName . '.edit', $id)
